working on birds #1

+++ GameTry/MovingPlatform.cpp
@@ -4,7 +4,7 @@ MovingPlatform::MovingPlatform(sf::Image & _image, sf::String _name, Level & _le
 	Entity(_image, _name, _x, _y, _spriteXInImage, _spriteYInImage, _spriteWInImage, _spriteHInImage)
 {
 	sprite.setTextureRect(sf::IntRect(spriteXInImage, spriteYInImage, spriteWInImage, spriteHInImage));
-	dx = 0.1;
+	dx = static_cast<float>(0.1);
 }
 
 void MovingPlatform::Update(float _time)
